+++ src/ScottPlot/Palettes/Dark.cs
@@ -1,0 +1,15 @@
+﻿/* This a qualitative 8-color palette generated using https://colorbrewer2.org
+ * © Cynthia Brewer, Mark Harrower and The Pennsylvania State University
+ * It is is both LCD and print friendly but not blind nor photocopy friendly
+ */
+namespace ScottPlot.Drawing.Colorsets
+{
+    class Dark : HexColorset, IPalette
+    {
+        public override string[] hexColors => new string[]
+        {
+            "#1b9e77","#d95f02","#7570b3","#e7298a","#66a61e",
+            "#e6ab02","#a6761d","#666666",
+        };
+    }
+}