+++ dev/changelog.md
@@ -5,6 +5,7 @@
 * Style: Refactored to use static classes instead of enumeration members (#1299, #1291)
 * NuGet: Improved System.Drawing.Common dependencies in user control packages (#1311, #1310) _Thanks @Kritner_
 * New Palettes:
+  * `Dark` 8-color palette LCD and print friendly (#1313) _Thanks @arthurits_
   * `DarkPastel` 8-color palette of LCD and print-friendly colors (#1313) _Thanks @arthurits_
 
 ## ScottPlot 4.1.23